platform/examples/pydantic-ai/requirements.txt
pydantic-ai[google]==1.0.15
python-dotenv==1.1.1
requests==2.32.4
